Adds tooltip to the vouch chip icons
