+++ web-app/package.json
@@ -41,6 +41,7 @@
     "react-dom": "^18.3.1",
     "react-qr-code": "^2.0.14",
     "react-router-dom": "^6.23.1",
+    "react-tooltip": "^5.28.0",
     "styled-components": "^6.1.11"
   },
   "devDependencies": {
